Delete in batches as an outbox could have more than 100 coalesced messages.

+++ src/sentry/models/outbox.py
@@ -596,15 +596,16 @@ class OutboxBase(Model):
             deleted_count = 0
 
             # Use a fetch and delete loop as doing cleanup in a single query
-            # causes timeouts with large datasets. Apply the ID condition in
-            # python as filtering rows in postgres leads to timeouts
-            delete_ids = []
-            for item in self.select_coalesced_messages()[:100]:
-                if item.id <= coalesced.id:
-                    delete_ids.append(item.id)
-
-            self.objects.filter(id__in=delete_ids).delete()
-            deleted_count = len(delete_ids)
+            # causes timeouts with large datasets. Fetch in batches of 100 and
+            # Apply the ID condition in python as filtering rows in postgres
+            # leads to timeouts.
+            while True:
+                batch = self.select_coalesced_messages()[:100]
+                delete_ids = [item.id for item in batch if item.id <= coalesced.id]
+                if not len(delete_ids):
+                    break
+                self.objects.filter(id__in=delete_ids).delete()
+                deleted_count += len(delete_ids)
 
             metrics.incr("outbox.processed", deleted_count, tags=tags)
             metrics.timing(
